+++ app/(root)/layout.tsx
@@ -4,7 +4,7 @@ import { Poppins } from 'next/font/google';
 
 import '../globals.css';
 import { Providers } from './providers';
-import NavigationBar from '@/components/NavigationBar';
+import NavigationBar from '@/components/navigation/NavigationBar';
 
 const poppins = Poppins({ weight: ['400', '600'], subsets: ['latin'] });
