+++ app/(root)/page.tsx
@@ -1,7 +1,7 @@
-import Hero from '@/components/Hero';
-import Skills from '@/components/Skills';
-import Services from '@/components/Services';
-import Work from '@/components/Work';
+import Hero from '@/components/hero/Hero';
+import Skills from '@/components/skills/Skills';
+import Services from '@/components/service/Services';
+import Work from '@/components/work/Work';
 import Project from '@/components/projects/Project';
 import Testimonials from '@/components/testimonials/Testimonials';
