+++ components/skills/SkillsCard.tsx
