+++ components/testimonials/ArrowRightControl.tsx
@@ -1,0 +1,21 @@
+import React from 'react';
+
+import ArrowRightIcon from './ArrowRightIcon';
+import { ArrowRightControlProps } from '@/types';
+
+const ArrowRightControl: React.FC<ArrowRightControlProps> = ({
+  onClick,
+  isHiddenOnLg = false,
+}) => {
+  let className =
+    'mx-[0.4rem] h-[3rem] w-[3rem] cursor-pointer flex items-center justify-center rounded-full bg-white-900 px-[1rem] dark:bg-black-200';
+  className += isHiddenOnLg ? ' lg:hidden' : ' hidden lg:flex';
+
+  return (
+    <div onClick={onClick} className={className}>
+      <ArrowRightIcon />
+    </div>
+  );
+};
+
+export default ArrowRightControl;
