+++ types/index.ts
@@ -1,3 +1,4 @@
+import React from 'react';
 import { PortableTextBlock } from 'sanity';
 
 export type ProfileType = {
@@ -64,4 +65,67 @@ export type ProjectCardContentProps = {
 export type ProjectCardImageProps = {
   image: string;
   heading: string;
+};
+
+export type Testimonial = {
+  _id: string;
+  headingName: string;
+  subheadingName: string;
+  content: string;
+  testimonialImage: string;
+  altText: string;
+};
+
+export type CarouselProps = {
+  testimonials: Testimonial[];
+};
+
+export type ArrowLeftControlProps = {
+  onClick: () => void;
+  isHiddenOnLg?: boolean;
+};
+
+export type ArrowRightControlProps = {
+  onClick: () => void;
+  isHiddenOnLg?: boolean;
+};
+
+export type StarComponentProps = {
+  count: number;
+};
+
+export type TestimonialDetailsProps = {
+  content: string;
+  headingName: string;
+  subheadingName: string;
+  show: boolean;
+};
+
+export interface WrapperDark200Props {
+  children: React.ReactNode;
+}
+
+export type Service = {
+  _id: string;
+  heading: string;
+  content: string;
+  whiteImage: string;
+  blueImage: string;
+};
+
+export type SkillsProps = {
+  _id: string;
+  image: string;
+  altText: string;
+};
+
+export interface ClipboardTextProps {
+  text: string;
+}
+
+export type WorkCardProps = {
+  title: string;
+  subtitle: string;
+  description: string;
+  imageSrc: string;
 };
